+++ module/Core/test/Exception/DeleteShortUrlExceptionTest.php
@@ -6,6 +6,7 @@ namespace ShlinkioTest\Shlink\Core\Exception;
 
 use PHPUnit\Framework\TestCase;
 use Shlinkio\Shlink\Core\Exception\DeleteShortUrlException;
+use Shlinkio\Shlink\Core\Model\ShortUrlIdentifier;
 
 use function Functional\map;
 use function range;
@@ -23,7 +24,10 @@ class DeleteShortUrlExceptionTest extends TestCase
         string $shortCode,
         string $expectedMessage,
     ): void {
-        $e = DeleteShortUrlException::fromVisitsThreshold($threshold, $shortCode);
+        $e = DeleteShortUrlException::fromVisitsThreshold(
+            $threshold,
+            ShortUrlIdentifier::fromShortCodeAndDomain($shortCode),
+        );
 
         self::assertEquals($threshold, $e->getVisitsThreshold());
         self::assertEquals($expectedMessage, $e->getMessage());
@@ -41,10 +45,29 @@ class DeleteShortUrlExceptionTest extends TestCase
     {
         return map(range(5, 50, 5), function (int $number) {
             return [$number, $shortCode = generateRandomShortCode(6), sprintf(
-                'Impossible to delete short URL with short code "%s" since it has more than "%s" visits.',
+                'Impossible to delete short URL with short code "%s", since it has more than "%s" visits.',
                 $shortCode,
                 $number,
             )];
         });
+    }
+
+    /** @test */
+    public function domainIsPartOfAdditionalWhenProvidedInIdentifier(): void
+    {
+        $e = DeleteShortUrlException::fromVisitsThreshold(
+            10,
+            ShortUrlIdentifier::fromShortCodeAndDomain('abc123', 'doma.in'),
+        );
+        $expectedMessage = 'Impossible to delete short URL with short code "abc123" for domain "doma.in", since it '
+            . 'has more than "10" visits.';
+
+        self::assertEquals([
+            'shortCode' => 'abc123',
+            'domain' => 'doma.in',
+            'threshold' => 10,
+        ], $e->getAdditionalData());
+        self::assertEquals($expectedMessage, $e->getMessage());
+        self::assertEquals($expectedMessage, $e->getDetail());
     }
 }